+++ src/ui/gui_views/btc_only/multi_sig/gui_import_multisig_wallet_info_view.c
@@ -4,7 +4,7 @@
 #include "gui_status_bar.h"
 #include "gui_import_multisig_wallet_info_widgets.h"
 
-int32_t GuiImportWalletInfoViewEventProcess(void *self, uint16_t usEvent, void *param, uint16_t usLen)
+int32_t GuiImportMultisigWalletInfoViewEventProcess(void *self, uint16_t usEvent, void *param, uint16_t usLen)
 {
     GUI_ASSERT(g_importMultisigWalletInfoView.isActive);
 
@@ -15,17 +15,8 @@ int32_t GuiImportWalletInfoViewEventProcess(void *self, uint16_t usEvent, void *
     case GUI_EVENT_OBJ_DEINIT:
         GuiImportMultisigWalletInfoWidgetsDeInit();
         break;
-    case GUI_EVENT_REFRESH:
-        GuiImportMultisigWalletInfoWidgetsRefresh();
-        break;
-    case GUI_EVENT_RESTART:
-        GuiImportMultisigWalletInfoWidgetsRestart();
-        break;
-    case SIG_SETUP_VIEW_TILE_PREV:
-        GuiImportMultiInfoPrevTile();
-        break;
-    case SIG_SETUP_VIEW_TILE_NEXT:
-        GuiImportMultiInfoNextTile();
+    case SIG_VERIFY_PASSWORD_PASS:
+        GuiImportMultisigWalletInfoVerifyPasswordSuccess();
         break;
     default:
         return ERR_GUI_UNHANDLED;
@@ -38,5 +29,5 @@ GUI_VIEW g_importMultisigWalletInfoView = {
     .previous = NULL,
     .isActive = false,
     .optimization = false,
-    .pEvtHandler = GuiImportWalletInfoViewEventProcess,
+    .pEvtHandler = GuiImportMultisigWalletInfoViewEventProcess,
 };
