+++ src/ui/gui_views/btc_only/multi_sig/gui_multisig_wallet_export_view.c
@@ -3,35 +3,25 @@
 #include "gui_resource.h"
 #include "gui_views.h"
 #include "gui_status_bar.h"
-#include "gui_import_multisig_wallet_widgets.h"
+#include "gui_multisig_wallet_export_widgets.h"
 
 int32_t GuiImportMultiSigWalletViewEventProcess(void *self, uint16_t usEvent, void *param, uint16_t usLen)
 {
-    GUI_ASSERT(g_importMultisigWalletView.isActive);
-    char* walletConifg;
+    GUI_ASSERT(g_multisigWalletExportView.isActive);
+    char* verifyCode;
     switch (usEvent) {
     case GUI_EVENT_OBJ_INIT:
         if (param == NULL) {
             return ERR_GUI_ERROR;
         }
-        walletConifg = (char *)param;
-        GuiImportMultisigWalletWidgetsInit(walletConifg, usLen);
+        verifyCode = (char *)param;
+        GuiMultisigWalletExportWidgetsInit(verifyCode, usLen);
         break;
     case GUI_EVENT_OBJ_DEINIT:
-        GuiImportMultisigWalletWidgetsDeInit();
+        GuiMultisigWalletExportWidgetsDeInit();
         break;
     case GUI_EVENT_REFRESH:
-        GuiImportMultisigWalletWidgetsRefresh();
-        break;
-    case SIG_SETUP_VIEW_TILE_PREV:
-        GuiImportMultiPrevTile();
-        break;
-    case SIG_SETUP_VIEW_TILE_NEXT:
-        GuiImportMultiNextTile();
-        break;
-    case SIG_VERIFY_PASSWORD_PASS:
-        GuiImportMultiNextTile();
-        GuiImportMultisigWalletInfoVerifyPasswordSuccess();
+        GuiMultisigWalletExportWidgetsRefresh();
         break;
     case SIG_BACKGROUND_UR_GENERATE_SUCCESS:
         GuiAnimantingQRCodeFirstUpdate((char*)param, usLen);
@@ -39,17 +29,14 @@ int32_t GuiImportMultiSigWalletViewEventProcess(void *self, uint16_t usEvent, vo
     case SIG_BACKGROUND_UR_UPDATE:
         GuiAnimatingQRCodeUpdate((char*)param, usLen);
         break;
-    case SIG_MULTISIG_WALLET_SET_WALLET_EXPORT:
-        GuiSetExportMultiSigSwitch();
-        break;
     default:
         return ERR_GUI_UNHANDLED;
     }
     return SUCCESS_CODE;
 }
 
-GUI_VIEW g_importMultisigWalletView = {
-    .id = SCREEN_MULTI_SIG_IMPORT_WALLET_SUCCESS,
+GUI_VIEW g_multisigWalletExportView = {
+    .id = SCREEN_MULTISIG_WALLET_EXPORT,
     .previous = NULL,
     .isActive = false,
     .optimization = false,
