+++ src/ui/gui_widgets/btc_only/multi_sig/gui_multisig_read_sdcard_widgets.c
@@ -1,0 +1,74 @@
+#include "gui_multisig_read_sdcard_widgets.h"
+#include "gui.h"
+#include "gui_page.h"
+#include "stdlib.h"
+#include "gui_button.h"
+#include "gui_import_multisig_wallet_info_widgets.h"
+
+#ifndef COMPILE_SIMULATOR
+#include "drv_sdcard.h"
+#include "user_fatfs.h"
+#else
+#include "simulator_mock_define.h"
+#include "simulator_model.h"
+#endif
+
+static PageWidget_t *g_pageWidget;
+static char g_fileList[10][64] = {0};
+
+static void GuiContent(lv_obj_t *);
+static void GuiSelectFileHandler(lv_event_t *e);
+
+void GuiMultisigReadSdcardWidgetsInit()
+{
+    g_pageWidget = CreatePageWidget();
+    GuiContent(g_pageWidget->contentZone);
+    SetNavBarLeftBtn(g_pageWidget->navBarWidget, NVS_BAR_RETURN, CloseCurrentViewHandler, NULL);
+    SetMidBtnLabel(g_pageWidget->navBarWidget, NVS_BAR_MID_LABEL, _("import_multi_wallet_via_micro_title"));
+    SetNavBarMidBtn(g_pageWidget->navBarWidget, NVS_MID_BUTTON_BUTT, NULL, NULL);
+}
+void GuiMultisigReadSdcardWidgetsDeInit()
+{
+    DestroyPageWidget(g_pageWidget);
+    g_pageWidget = NULL;
+}
+
+static void GuiContent(lv_obj_t *parent)
+{
+    GuiAddObjFlag(parent, LV_OBJ_FLAG_SCROLLABLE);
+    char *buffer = EXT_MALLOC(1024 * 5);
+    uint32_t number = 0;
+    int i = 0;
+#ifdef COMPILE_SIMULATOR
+    FatfsGetFileName("C:/assets/sd", buffer, &number, 1024 * 5);
+#else
+    FatfsGetFileName("0:", buffer, &number, 1024 * 5);
+#endif
+    char *token = strtok(buffer, " ");
+    while (token != NULL)
+    {
+        strncpy(g_fileList[i], token, sizeof(g_fileList[i]));
+        token = strtok(NULL, " ");
+        lv_obj_t *btn = GuiCreateSelectButton(parent, g_fileList[i], &imgArrowRight, GuiSelectFileHandler, g_fileList[i], false);
+        lv_obj_align(btn, LV_ALIGN_TOP_MID, 0, 84 * i);
+        i++;
+        if (i == 10)
+        {
+            break;
+        }
+    }
+    EXT_FREE(buffer);
+}
+
+static void GuiSelectFileHandler(lv_event_t *e)
+{
+    lv_event_code_t code = lv_event_get_code(e);
+    char *path = lv_event_get_user_data(e);
+
+    if (code == LV_EVENT_CLICKED)
+    {
+        char *walletConfig = FatfsFileRead(path);
+        GuiSetMultisigImportWalletDataBySDCard(walletConfig);
+        GuiFrameOpenView(&g_importMultisigWalletInfoView);
+    }
+}
