+++ src/ui/gui_widgets/btc_only/multi_sig/gui_multisig_read_sdcard_widgets.h
@@ -1,0 +1,7 @@
+#ifndef _GUI_MULTISIG_READ_SDCARD_WIDGETS_H_
+#define _GUI_MULTISIG_READ_SDCARD_WIDGETS_H_
+
+void GuiMultisigReadSdcardWidgetsInit();
+void GuiMultisigReadSdcardWidgetsDeInit();
+
+#endif
