+++ src/ui/gui_widgets/btc_only/multi_sig/gui_multisig_select_import_method_widgets.h
@@ -1,0 +1,7 @@
+#ifndef _GUI_MULTISIG_SELECT_IMPORT_METHOD_WIDGETS_H_
+#define _GUI_MULTISIG_SELECT_IMPORT_METHOD_WIDGETS_H_
+
+void GuiMultisigSelectImportMethodWidgetsInit();
+void GuiMultisigSelectImportMethodWidgetsDeInit();
+
+#endif
